feat: Add 'documenten.lezen' and 'zaken.lezen' scopes (#179)
* feat: Add 'documenten.lezen' and 'zaken.lezen' scopes

This commit introduces two new scopes, 'documenten.lezen' and 'zaken.lezen', to the catalogi-api. These scopes inherit the same permissions as the existing 'catalogi.lezen' scope. The addition of these scopes is necessary to accommodate the new expand model, which involves requesting information from the catalogi-api through the documenten-api and zaken-api.

Issue #2262